refactor(core): split sessions manager actor in different files
